Transferred legacy tests from test_load_accucor_msruns.py and improved ArchiveFile.objects.get_or_create based on the transferred tests.
Details:

- Transferred accucor loader tests to the archive file model tests.  (Methods in the accucor loader had been transferred to ArchiveFile.
- Transferred accucor loader tests to the load_msruns tests.  (The accucor loader created MSRunSample and MSRunSequence records, both of which are done in separate loaders now.)
- ArchiveFile.objects.get_or_create now raises a ValueError if there was no file supplied AND no checksum is provided.
- Moved the DataType and DataFormat conversions from string to record ID to the top of the get_or_create method, due to the placement of the call to super().get_or_create which expects an ID.

Files checked in: DataRepo/models/archive_file.py DataRepo/tests/management/commands/legacy/test_load_accucor_msruns.py DataRepo/tests/management/commands/test_load_msruns.py DataRepo/tests/models/test_archive_file.py
